+++ tests/e2e/data_persistence.go
@@ -576,10 +576,3 @@ var _ = ginkgo.Describe("Data Persistence", func() {
 		deleteResourceQuota(client, namespace)
 	})
 })
-
-func createAndVerifyFilesOnVolume(namespace string, podname string,
-	newEmptyfilesToCreate []string, filesToCheck []string) {
-	createEmptyFilesOnVSphereVolume(namespace, podname, newEmptyfilesToCreate)
-	ginkgo.By(fmt.Sprintf("Verify files exist on volume mounted on: %v", podname))
-	verifyFilesExistOnVSphereVolume(namespace, podname, poll, pollTimeoutShort, filesToCheck...)
-}
