+++ tests/e2e/volume_expansion_utils.go
@@ -17,14 +17,19 @@ limitations under the License.
 package e2e
 
 import (
+	"context"
 	"fmt"
 	"strconv"
 	"strings"
+	"sync"
+	"time"
 
 	"github.com/onsi/ginkgo/v2"
 	"github.com/onsi/gomega"
 	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/util/wait"
 	clientset "k8s.io/client-go/kubernetes"
 	"k8s.io/kubernetes/test/e2e/framework"
 	e2ekubectl "k8s.io/kubernetes/test/e2e/framework/kubectl"
@@ -103,4 +108,202 @@ func increaseSizeOfPvcAttachedToPodWithoutF(client clientset.Interface,
 	gomega.Expect(fsSize).Should(gomega.BeNumerically(">", originalSizeInMb),
 		fmt.Sprintf("error updating filesystem size for %q. Resulting filesystem size is %d", pvclaim.Name, fsSize))
 	ginkgo.By("File system resize finished successfully")
+}
+
+// resize this method does PVC volume expansion to given size
+func resize(client clientset.Interface, pvc *v1.PersistentVolumeClaim,
+	currentPvcSize resource.Quantity, newSize resource.Quantity, wg *sync.WaitGroup) {
+	defer wg.Done()
+	framework.Logf("currentPvcSize %v, newSize %v", currentPvcSize, newSize)
+	_, err := expandPVCSize(pvc, newSize, client)
+	time.Sleep(30 * time.Second)
+	framework.Logf("Error from expansion attempt on pvc %v, to %v: %v", pvc.Name, newSize, err)
+}
+
+// verifyPVCRequestedSizeInSupervisor compares
+// Pvc.Spec.Resources.Requests[v1.ResourceStorage] with given size in SVC.
+func verifyPVCRequestedSizeInSupervisor(pvcName string, size resource.Quantity) bool {
+	SvcPvc := getPVCFromSupervisorCluster(pvcName)
+	pvcSize := SvcPvc.Spec.Resources.Requests[v1.ResourceStorage]
+	framework.Logf("SVC PVC requested size: %v", pvcSize)
+	return pvcSize.Cmp(size) == 0
+}
+
+// verifyPvcRequestedSizeUpdateInSupervisorWithWait waits till
+// Pvc.Spec.Resources.Requests[v1.ResourceStorage] matches with given size
+// in SVC.
+func verifyPvcRequestedSizeUpdateInSupervisorWithWait(pvcName string, size resource.Quantity) (bool, error) {
+	var b bool
+	waitErr := wait.PollUntilContextTimeout(context.Background(), resizePollInterval, totalResizeWaitPeriod, true,
+		func(ctx context.Context) (bool, error) {
+			b = verifyPVCRequestedSizeInSupervisor(pvcName, size)
+			return b, nil
+		})
+	return b, waitErr
+}
+
+// waitForSvcPvcToReachFileSystemResizePendingCondition waits for SVC PVC to
+// reach FileSystemResizePendingCondition status condition.
+func waitForSvcPvcToReachFileSystemResizePendingCondition(ctx context.Context, svcPvcName string,
+	timeout time.Duration) error {
+	waitErr := wait.PollUntilContextTimeout(ctx, resizePollInterval, timeout, true,
+		func(ctx context.Context) (bool, error) {
+			_, err := checkSvcPvcHasGivenStatusCondition(svcPvcName, true, v1.PersistentVolumeClaimFileSystemResizePending)
+			if err == nil {
+				return true, nil
+			}
+			if strings.Contains(err.Error(), "not matching") {
+				return false, nil
+			}
+			return false, err
+		})
+	return waitErr
+}
+
+// This method verifies PVC size after expansion with polling time
+func verifyPVSizeinSupervisorWithWait(ctx context.Context, svcPVCName string, newSize resource.Quantity) error {
+	waitErr := wait.PollUntilContextTimeout(ctx, resizePollInterval, totalResizeWaitPeriod, true,
+		func(ctx context.Context) (bool, error) {
+			svcPV := getPvFromSupervisorCluster(svcPVCName)
+			svcPVSize := svcPV.Spec.Capacity[v1.ResourceStorage]
+			return svcPVSize.Cmp(newSize) >= 0, nil
+		})
+	return waitErr
+}
+
+// waitForPVCToReachFileSystemResizePendingCondition waits for PVC to reach
+// FileSystemResizePendingCondition status condition.
+func waitForPVCToReachFileSystemResizePendingCondition(client clientset.Interface,
+	namespace string, pvcName string, timeout time.Duration) (*v1.PersistentVolumeClaim, error) {
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+	var pvclaim *v1.PersistentVolumeClaim
+	var err error
+
+	waitErr := wait.PollUntilContextTimeout(ctx, resizePollInterval, timeout, true,
+		func(ctx context.Context) (bool, error) {
+			pvclaim, err = client.CoreV1().PersistentVolumeClaims(namespace).Get(ctx, pvcName, metav1.GetOptions{})
+			if err != nil {
+				return false, fmt.Errorf("error while fetching pvc '%v' after controller resize: %v", pvcName, err)
+			}
+			gomega.Expect(pvclaim).NotTo(gomega.BeNil())
+
+			inProgressConditions := pvclaim.Status.Conditions
+			// If there are conditions on the PVC, it must be of
+			// FileSystemResizePending type.
+			if len(inProgressConditions) > 0 {
+				expectEqual(len(inProgressConditions), 1, fmt.Sprintf("PVC '%v' has more than one status conditions", pvcName))
+				if inProgressConditions[0].Type == v1.PersistentVolumeClaimFileSystemResizePending {
+					return true, nil
+				}
+				gomega.Expect(inProgressConditions[0].Type == v1.PersistentVolumeClaimResizing).To(gomega.BeTrue(),
+					fmt.Sprintf("PVC '%v' is not in 'Resizing' or 'FileSystemResizePending' status condition", pvcName))
+			} else {
+				return false, fmt.Errorf(
+					"resize was not triggered on PVC '%v' or no status conditions related to resizing found on it", pvcName)
+			}
+			return false, nil
+		})
+	return pvclaim, waitErr
+}
+
+// checkPvcHasGivenStatusCondition checks if the status condition in PVC
+// matches with the one we want.
+func checkPvcHasGivenStatusCondition(client clientset.Interface, namespace string, pvcName string,
+	conditionsPresent bool, condition v1.PersistentVolumeClaimConditionType) (*v1.PersistentVolumeClaim, error) {
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+	pvclaim, err := client.CoreV1().PersistentVolumeClaims(namespace).Get(ctx, pvcName, metav1.GetOptions{})
+	gomega.Expect(err).NotTo(gomega.HaveOccurred())
+	gomega.Expect(pvclaim).NotTo(gomega.BeNil())
+	inProgressConditions := pvclaim.Status.Conditions
+
+	if len(inProgressConditions) == 0 {
+		if conditionsPresent {
+			return pvclaim, fmt.Errorf("no status conditions found on PVC: %v", pvcName)
+		}
+		return pvclaim, nil
+	}
+	expectEqual(len(inProgressConditions), 1, fmt.Sprintf("PVC '%v' has more than one status condition", pvcName))
+	if inProgressConditions[0].Type != condition {
+		return pvclaim, fmt.Errorf(
+			"status condition found on PVC '%v' is '%v', and is not matching with expected status condition '%v'",
+			pvcName, inProgressConditions[0].Type, condition)
+	}
+	return pvclaim, nil
+}
+
+// waitForFSResize waits for the filesystem in the pv to be resized.
+func waitForFSResizeInSvc(svcPVCName string) (*v1.PersistentVolumeClaim, error) {
+	svcClient, _ := getSvcClientAndNamespace()
+	svcPvc := getPVCFromSupervisorCluster(svcPVCName)
+	return waitForFSResize(svcPvc, svcClient)
+}
+
+// onlineVolumeResizeCheck this method increases the PVC size, which is attached
+// to POD.
+func onlineVolumeResizeCheck(f *framework.Framework, client clientset.Interface,
+	namespace string, svcPVCName string, volHandle string, pvclaim *v1.PersistentVolumeClaim, pod *v1.Pod) {
+	var originalSizeInMb int64
+	var err error
+	// Fetch original FileSystemSize.
+	ginkgo.By("Verify filesystem size for mount point /mnt/volume1 before expansion")
+	originalSizeInMb, err = getFileSystemSizeForOsType(f, client, pod)
+	gomega.Expect(err).NotTo(gomega.HaveOccurred())
+
+	// Resize PVC.
+	// Modify PVC spec to trigger volume expansion.
+	ginkgo.By("Expanding current pvc")
+	currentPvcSize := pvclaim.Spec.Resources.Requests[v1.ResourceStorage]
+	newSize := currentPvcSize.DeepCopy()
+	newSize.Add(resource.MustParse("1Gi"))
+	framework.Logf("currentPvcSize %v, newSize %v", currentPvcSize, newSize)
+	pvclaim, err = expandPVCSize(pvclaim, newSize, client)
+	gomega.Expect(pvclaim).NotTo(gomega.BeNil())
+	gomega.Expect(err).NotTo(gomega.HaveOccurred())
+
+	_, err = waitForFSResizeInSvc(svcPVCName)
+	gomega.Expect(err).NotTo(gomega.HaveOccurred())
+
+	ginkgo.By("Waiting for file system resize to finish")
+	pvclaim, err = waitForFSResize(pvclaim, client)
+	framework.ExpectNoError(err, "while waiting for fs resize to finish")
+
+	pvcConditions := pvclaim.Status.Conditions
+	expectEqual(len(pvcConditions), 0, "pvc should not have conditions")
+
+	ginkgo.By(fmt.Sprintf("Invoking QueryCNSVolumeWithResult with VolumeID: %s", volHandle))
+	queryResult, err := e2eVSphere.queryCNSVolumeWithResult(volHandle)
+	gomega.Expect(err).NotTo(gomega.HaveOccurred())
+
+	if len(queryResult.Volumes) == 0 {
+		err = fmt.Errorf("QueryCNSVolumeWithResult returned no volume")
+	}
+	gomega.Expect(err).NotTo(gomega.HaveOccurred())
+
+	var fsSize int64
+	ginkgo.By("Verify filesystem size for mount point /mnt/volume1")
+	fsSize, err = getFileSystemSizeForOsType(f, client, pod)
+	gomega.Expect(err).NotTo(gomega.HaveOccurred())
+	framework.Logf("File system size after expansion : %d", fsSize)
+	// Filesystem size may be smaller than the size of the block volume
+	// so here we are checking if the new filesystem size is greater than
+	// the original volume size as the filesystem is formatted for the
+	// first time.
+	gomega.Expect(fsSize).Should(gomega.BeNumerically(">", originalSizeInMb),
+		fmt.Sprintf("error updating filesystem size for %q. Resulting filesystem size is %d", pvclaim.Name, fsSize))
+	ginkgo.By("File system resize finished successfully")
+
+	framework.Logf("Online volume expansion in GC PVC is successful")
+
+}
+
+// verifyPVSizeinSupervisor compares PV.Spec.Capacity[v1.ResourceStorage] in
+// SVC with the size passed here.
+func verifyPVSizeinSupervisor(svcPVCName string, newSize resource.Quantity) {
+	svcPV := getPvFromSupervisorCluster(svcPVCName)
+	svcPVSize := svcPV.Spec.Capacity[v1.ResourceStorage]
+	// If pv size is greater or equal to requested size that means controller
+	// resize is finished.
+	gomega.Expect(svcPVSize.Cmp(newSize) >= 0).To(gomega.BeTrue())
 }
